+++ src/components/GenericPagesComponents/ButtonLinkNav/ButtonLinkNav.tsx
@@ -1,8 +1,8 @@
 import { FC } from "react";
 
-import ButtonAsLink from "../SharedComponents/Button/ButtonAsLink";
-import useIsCurrent from "../MenuLinks/useIsCurrent";
-import { HTMLAnchorProps } from "../SharedComponents/Button/common";
+import ButtonAsLink from "../../SharedComponents/Button/ButtonAsLink";
+import useIsCurrent from "../../MenuLinks/useIsCurrent";
+import { HTMLAnchorProps } from "../../SharedComponents/Button/common";
 
 import Flex, { FlexProps } from "@/components/SharedComponents/Flex";
 import Box from "@/components/SharedComponents/Box";
